requirements.txt
altair                    5.5.0
attrs                     25.4.0
blinker                   1.9.0
cachetools                6.2.1
certifi                   2025.10.5
charset-normalizer        3.4.4
click                     8.3.0
colorama                  0.4.6
comtypes                  1.4.13
contourpy                 1.3.3
cycler                    0.12.1
filelock                  3.20.0
filetype                  1.2.0
filterpy                  1.4.5
fonttools                 4.60.1
fsspec                    2025.9.0
gitdb                     4.0.12
GitPython                 3.1.45
idna                      3.7
Jinja2                    3.1.6
joblib                    1.5.2
jsonschema                4.25.1
jsonschema-specifications 2025.9.1
kiwisolver                1.4.9
lap                       0.5.12
MarkupSafe                3.0.3
matplotlib                3.10.7
mpmath                    1.3.0
narwhals                  2.10.0
networkx                  3.5
numpy                     2.2.6
opencv-python             4.10.0.84
packaging                 25.0
pandas                    2.3.3
pi_heif                   1.1.1
pillow                    11.3.0
pillow-avif-plugin        1.5.2
pip                       25.3
playsound                 1.2.2
protobuf                  6.33.0
psutil                    7.1.2
py-cpuinfo                9.0.0
pyarrow                   22.0.0
pydeck                    0.9.1
pyparsing                 3.2.5
pypiwin32                 223
python-dateutil           2.9.0.post0
python-dotenv             1.2.1
pyttsx3                   2.90
pytz                      2025.2
pywin32                   311
PyYAML                    6.0.3
referencing               0.37.0
requests                  2.32.5
requests-toolbelt         1.0.0
roboflow                  1.2.11
rpds-py                   0.28.0
scikit-learn              1.7.2
scipy                     1.16.2
seaborn                   0.13.2
setuptools                80.9.0
six                       1.17.0
smmap                     5.0.2
streamlit                 1.50.0
sympy                     1.14.0
tenacity                  9.1.2
threadpoolctl             3.6.0
toml                      0.10.2
torch                     2.9.0
torchvision               0.24.0
tornado                   6.5.2
tqdm                      4.67.1
typing_extensions         4.15.0
tzdata                    2025.2
ultralytics               8.3.12
ultralytics-thop          2.0.17
urllib3                   2.5.0
watchdog                  6.0.0